market-pulse/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
httpx==0.28.1
yfinance==0.2.51
pandas==2.2.3
fredapi==0.5.2
apscheduler==3.10.4
python-dotenv==1.0.1
